chore: implement forward refs
