+++ packages/components/stepper/src/step-description.tsx
@@ -1,12 +1,20 @@
-import { chakra } from "@chakra-ui/system"
+import { HTMLChakraProps, chakra, forwardRef } from "@chakra-ui/system"
 import { useStepContext, useStepperStyles } from "./step-context"
 
-export type StepDescriptionProps = {
-  children: React.ReactNode
-}
+export type StepDescriptionProps = HTMLChakraProps<"p">
 
-export function StepDescription(props: StepDescriptionProps) {
+export const StepDescription = forwardRef<{}, "p">(function StepDescription(
+  props: StepDescriptionProps,
+  ref,
+) {
   const { status } = useStepContext()
   const styles = useStepperStyles()
-  return <chakra.p data-status={status} {...props} __css={styles.description} />
-}
+  return (
+    <chakra.p
+      ref={ref}
+      data-status={status}
+      {...props}
+      __css={styles.description}
+    />
+  )
+})
