+++ packages/components/stepper/src/step-title.tsx
@@ -1,10 +1,15 @@
-import { HTMLChakraProps, chakra } from "@chakra-ui/system"
+import { HTMLChakraProps, chakra, forwardRef } from "@chakra-ui/system"
 import { useStepContext, useStepperStyles } from "./step-context"
 
-export type StepTitleProps = HTMLChakraProps<"h5">
+export type StepTitleProps = HTMLChakraProps<"h3">
 
-export function StepTitle(props: StepTitleProps) {
+export const StepTitle = forwardRef(function StepTitle(
+  props: StepTitleProps,
+  ref,
+) {
   const { status } = useStepContext()
   const styles = useStepperStyles()
-  return <chakra.h5 data-status={status} {...props} __css={styles.title} />
-}
+  return (
+    <chakra.h3 ref={ref} data-status={status} {...props} __css={styles.title} />
+  )
+})
